hapi errors

+++ src/services/superhero_v2.service.js
@@ -1,4 +1,5 @@
 const superheroModel = require('../models/superhero_v2.model');
+const Boom = require('@hapi/boom')
 
 class superHeroService {
   /* Promesas y funciones asincrónicas
@@ -13,6 +14,15 @@ class superHeroService {
     return superheroModel.find()
   }
 
+  /* Función que nos devuelve una promesa */
+  find(){
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve(superheroModel.find())
+      }, 3000)
+    })
+  }
+
   async showSuperHero(superheroId) {
     return superheroModel.findById({ _id: superheroId })
   }
@@ -21,7 +31,7 @@ class superHeroService {
     return superheroModel
       .findById({ _id: superheroId }, {superHero, realName, superPower})
       .then((superheroFind) => {
-        if (!superheroFind) throw Error('No se encontró el superheroe')
+        if (!superheroFind) throw Boom.notFound('No se encontró el superheroe')
         return superheroMoodel.updateOne(
           { superheroId },
           { superHero, realName, superPower }
@@ -31,6 +41,7 @@ class superHeroService {
 
   async removeSuperHero(superheroId) {
     const superhero_remove = superHeroModel.findById({ _id: superheroId })
+    if (!superheroFind) throw Boom.notFound('No se encontró el superheroe')
     superheroModel.deleteOne(superhero_remove)
   }
 }
